back-end/requirements.txt
Flask==2.0.2
Flask-Cors==3.0.10
mysql-connector-python==8.0.26
python-dotenv==0.19.2
